LayoutTests/media/modern-media-controls/pip-placard/pip-placard.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/placard.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-placard.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>PiPPlacard</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

debug("Placard with all parameters");
const placard = new PiPPlacard({ layoutTraits: LayoutTraits.macOS });

shouldBe("placard.children[0].children[0].iconName", "Icons.PiPPlacard");
shouldBeEqualToString("placard.children[0].children[1].element.innerText", "This video is playing in Picture in Picture");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
